Validated wt_downscaling and other related subroutines including find_the_days.c.

+++ src/libs/clim/clim.h
@@ -31,8 +31,8 @@
 
 void clim_daily_tserie_climyear(double *bufout, double *bufin, tstruct *buftime, double missing_val, int ni, int nj, int ntime);
 void remove_seasonal_cycle(double *bufout, double *clim, double *bufin, tstruct *buftime, double missing_val,
-                           int filter_width, char *type, short int clim_provided, int ni, int nj, int ntime);
-short int dayofclimyear(short int day, short int month);
+                           int filter_width, char *type, int clim_provided, int ni, int nj, int ntime);
+int dayofclimyear(int day, int month);
 
 #endif
